+++ src/pages/TotalSalesListPage/TotalSalesListPage.scss
@@ -1,0 +1,18 @@
+#sell-list{
+  text-align: center;
+  .ant-table-cell{
+    text-align: center;
+  }
+  .listNumber {
+    width: 40px;
+  }
+  .sellListItem {
+    width: 300px;
+  }
+  .totalPrice{
+    width: 100px;
+  }
+  .userBank{
+    width: 170px;
+  }
+}
